Complete about_scoring_project
I appreciate this is probably a really horrible solution for this. I
tried to reduce the if/else conditionals here to 1 line, a la the C ?
operator, but python's equivalent didn't seem to like it. Even given
that, there are probably all sorts of better solutions involving
python idioms I'm not yet aware of.